# opt: cache restricted interesting orderings