--- a/pkg/sql/opt/xform/testdata/rules/cycle
+++ b/pkg/sql/opt/xform/testdata/rules/cycle
@@ -62,7 +62,7 @@ expropt disable=MemoCycleTestRelRuleFilter skip-race
 ----
 error: memo group optimization passes surpassed limit of 100000; there may be a cycle in the memo
 details:
-memo (not optimized, ~15KB, required=[], cycle=[G1->G3->G5->G5])
+memo (not optimized, ~16KB, required=[], cycle=[G1->G3->G5->G5])
  ├── G1: (left-join G2 G3 G4)
  ├── G2: (scan ab,cols=()) (scan ab@ab_b_idx,cols=())
  │    └── []