… pkg/sql/opt/xform/testdata/rules/join_order
… pkg/sql/opt/xform/testdata/rules/limit
… pkg/sql/opt/xform/testdata/rules/select