#### opt: cache simplified interesting orderings

The `GenerateMergeJoins` rule simplifies the interesting orderings of a
join's left child with respect to the child's functional dependencies.
Simplifying an ordering set is computationally expensive, and it can
dominate the CPU-time of optimizing queries with many joins and
interesting orderings. In a query with many joins, `GenerateMergeJoins`
fires repeatedly with the same left child. To avoid repeatedly computing
the same simplified interesting orderings, we cache them in
`props.Relational.Rule.SimplifiedInterestingOrderings`. This
significantly reduces the planning time of some queries.

Release note (performance improvement): Query planning time has been reduced
for some queries with multiple joins.

#### opt: cache restricted interesting orderings

This commit builds on the last commit by caching interesting orderings
that are restricted to a specific column set, not just simplified
interesting orderings. Restricting an ordering set is computationally
expensive (see `OrderingSet.RestrictToCols`), and queries with many
joins can repeatedly restrict the same interesting orderings of a
relation with the same set of columns. Caching restricted orderings
eliminates this duplicate work.

Epic: None

Release note: None